Fixed error for merge_voices
Former-commit-id: f0abf890057d3ec14f4f7222b695c2f3eb0f0bda

+++ musif/process/utils.py
@@ -149,7 +149,7 @@ def _join_double_bass(df: DataFrame):
             df[col] = df[col].astype(str)
             df[formatted_col] = df[[col, formatted_col]].sum(axis=1)
             df[formatted_col] = [i.replace("nan", "") for i in df[formatted_col]]
-            df[formatted_col] = df[formatted_col].astype(float)
+            # df[formatted_col] = df[formatted_col].astype(float)
         else:
             df[col] = df[col].astype(float)
             df[formatted_col] = df[[formatted_col, col]].sum(axis=1)
